added logging

+++ EnvironmentSimulator/src/main/java/chainOfResponsibility/commandHandler/DeleteAnimalCommandHandler.java
@@ -1,22 +1,56 @@
 package chainOfResponsibility.commandHandler;
 
 import factoryMethod.AnimalFactory.AnimalNotFoundException;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 public class DeleteAnimalCommandHandler extends CommandHandler {
+
+    private static final Logger logger = LogManager.getLogger(DeleteAnimalCommandHandler.class);
+
     @Override
     public boolean handle(String cmd, java.util.Scanner scanner, template.Game.GameLoop gameLoop) {
+
+        // Global cancel support
+        if (isCancel(cmd)) {
+            System.out.println("Operation cancelled.");
+            logger.info("DeleteAnimal operation cancelled by user.");
+            return true;
+        }
+
         if (cmd.equalsIgnoreCase("deleteAnimal")) {
-            System.out.println("Enter the ID of the animal to delete:");
+
+            logger.info("DeleteAnimal command received. Requesting animal ID from user.");
+
+            System.out.println("Enter the ID of the animal to delete (or 'cancel' to cancel):");
             String idInput = scanner.hasNextLine() ? scanner.nextLine().trim() : "";
-            try {
-                gameLoop.animalRepository.remove(idInput);
-                System.out.println("Animal with ID " + idInput + " has been deleted.");
+
+            // Cancel inside the handler
+            if (checkCancel(idInput)) {
                 return true;
-            } catch (Exception e) {
-                throw new AnimalNotFoundException(idInput, gameLoop.animalRepository);
+            }
+
+            logger.debug("User entered ID: '{}'", idInput);
+
+            if (askAreYouSure(scanner)) {
+                try {
+                    gameLoop.animalRepository.remove(idInput);
+
+                    logger.info("Animal with ID '{}' successfully deleted.", idInput);
+                    System.out.println("Animal with ID " + idInput + " has been deleted.");
+                    return true;
+
+                } catch (Exception e) {
+
+                    logger.error("Failed to delete animal with ID '{}'. Throwing AnimalNotFoundException.", idInput);
+                    throw new AnimalNotFoundException(idInput, gameLoop.animalRepository);
+                }
             }
         }
-        return next != null && next.handle(cmd, scanner, gameLoop);
 
+        logger.trace("Command '{}' not handled by {}. Passing to next handler.",
+                cmd, this.getClass().getSimpleName());
+
+        return next != null && next.handle(cmd, scanner, gameLoop);
     }
 }
